dmsetup: allow empty cookie value for dmsetup udevflags/udevvars
Alleviate "dmsetup udevflags/udevvars --flags cookie" value check so that
empty value is allowed in which case there is no output. This will make the
udev rule calling the command a bit easier so it doesn't need to check for
empty cookie value itself. This applies mainly to new "dmsetup udevvars"
which can still output the other parts (--attrs --splitname) while ignoring
the --flags part if cookie is empty.

Then, we don't need to have two separate udev rules, one for the case when
cookie is present (genuine DM uevent) and the other one for the case when
cookie is not present (synthetic or any other uevents)

Signed-off-by: Peter Rajnoha <prajnoha@redhat.com>